test-r-multicast-sender: Improve macro's so they will react correctly on overflow
20070607153419-93b9a-fc3a677ba8170c38608f8da073b66b42c43dad17.gz

+++ tests/test-r-multicast-sender.c
@@ -8,9 +8,9 @@
 
 #define REPAIR_PACKET 15
 
-#define EXTRA_SEEN 11
-#define NR_PACKETS 40
-#define SERIAL_OFFSET (~0 - NR_PACKETS/2)
+#define EXTRA_SEEN ((guint32)11)
+#define NR_PACKETS ((guint32)40)
+#define SERIAL_OFFSET ((guint32)(~0 - NR_PACKETS/2))
 
 typedef struct {
   const gchar *id;
